Moved some code into separate manager
